[BinaryPlatforms]: Flesh out CPUID support
This PR fleshes out (and slightly renames) some pieces of the CPUID
machinery to better support BB's multi-march system.  It declares at
least one valid microarchitecture for all architectures, and splits the
CPUID module out into its own source file for readability.
