+++ base/binaryplatforms.jl
@@ -8,71 +8,7 @@ export AbstractPlatform, Platform, HostPlatform, platform_dlext, tags, arch, os,
 import .Libc.Libdl
 
 ### Submodule with information about CPU features
-module CPUID
-
-export cpu_isa
-
-"""
-    ISA(features::Set{UInt32})
-
-A structure which represents the Instruction Set Architecture (ISA) of a
-computer.  It holds the `Set` of features of the CPU.
-
-The numerical values of the features are automatically generated from the C
-source code of Julia and stored in the `features_h.jl` Julia file.
-"""
-struct ISA
-    features::Set{UInt32}
-end
-
-Base.:<=(a::ISA, b::ISA) = a.features <= b.features
-Base.:<(a::ISA,  b::ISA) = a.features <  b.features
-Base.isless(a::ISA,  b::ISA) = a < b
-
-include(string(length(Core.ARGS) >= 2 ? Core.ARGS[2] : "", "features_h.jl"))  # include($BUILDROOT/base/features_h.jl)
-
-# Keep in sync with `arch_march_isa_mapping`.
-const ISAs_by_family = Dict(
-    "x86_64" => [
-        # Source: https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html.
-        # Implicit in all sets, because always required: mmx, sse, sse2
-        "x86_64" => ISA(Set{UInt32}()),
-        "core2" => ISA(Set((JL_X86_sse3, JL_X86_ssse3))),
-        "nehalem" => ISA(Set((JL_X86_sse3, JL_X86_ssse3, JL_X86_sse41, JL_X86_sse42, JL_X86_popcnt))),
-        "sandybridge" => ISA(Set((JL_X86_sse3, JL_X86_ssse3, JL_X86_sse41, JL_X86_sse42, JL_X86_popcnt, JL_X86_avx, JL_X86_aes, JL_X86_pclmul))),
-        "haswell" => ISA(Set((JL_X86_movbe, JL_X86_sse3, JL_X86_ssse3, JL_X86_sse41, JL_X86_sse42, JL_X86_popcnt, JL_X86_avx, JL_X86_avx2, JL_X86_aes, JL_X86_pclmul, JL_X86_fsgsbase, JL_X86_rdrnd, JL_X86_fma, JL_X86_bmi, JL_X86_bmi2, JL_X86_f16c))),
-        "skylake" => ISA(Set((JL_X86_movbe, JL_X86_sse3, JL_X86_ssse3, JL_X86_sse41, JL_X86_sse42, JL_X86_popcnt, JL_X86_avx, JL_X86_avx2, JL_X86_aes, JL_X86_pclmul, JL_X86_fsgsbase, JL_X86_rdrnd, JL_X86_fma, JL_X86_bmi, JL_X86_bmi2, JL_X86_f16c, JL_X86_rdseed, JL_X86_adx, JL_X86_prfchw, JL_X86_clflushopt, JL_X86_xsavec, JL_X86_xsaves))),
-        "skylake_avx512" => ISA(Set((JL_X86_movbe, JL_X86_sse3, JL_X86_ssse3, JL_X86_sse41, JL_X86_sse42, JL_X86_popcnt, JL_X86_pku, JL_X86_avx, JL_X86_avx2, JL_X86_aes, JL_X86_pclmul, JL_X86_fsgsbase, JL_X86_rdrnd, JL_X86_fma, JL_X86_bmi, JL_X86_bmi2, JL_X86_f16c, JL_X86_rdseed, JL_X86_adx, JL_X86_prfchw, JL_X86_clflushopt, JL_X86_xsavec, JL_X86_xsaves, JL_X86_avx512f, JL_X86_clwb, JL_X86_avx512vl, JL_X86_avx512bw, JL_X86_avx512dq, JL_X86_avx512cd))),
-    ],
-    "arm" => [
-        "armv7l" => ISA(Set{UInt32}()),
-        "armv7l_neon" => ISA(Set((JL_AArch32_neon,))),
-        "armv7l_neon_vfp4" => ISA(Set((JL_AArch32_neon, JL_AArch32_vfp4))),
-    ],
-    "aarch64" => [
-        # Implicit in all sets, because always required: fp, asimd
-        "armv8.0_a" => ISA(Set{UInt32}()),
-        "armv8.1_a" => ISA(Set((JL_AArch64_lse, JL_AArch64_crc, JL_AArch64_rdm))),
-        "armv8.2_a_crypto" => ISA(Set((JL_AArch64_lse, JL_AArch64_crc, JL_AArch64_rdm, JL_AArch64_aes, JL_AArch64_sha2))),
-        "armv8.4_a_crypto_sve" => ISA(Set((JL_AArch64_lse, JL_AArch64_crc, JL_AArch64_rdm, JL_AArch64_fp16fml, JL_AArch64_dotprod, JL_AArch64_aes, JL_AArch64_sha2, JL_AArch64_dotprod, JL_AArch64_sve))),
-    ],
-)
-
-test_cpu_feature(feature::UInt32) = ccall(:jl_test_cpu_feature, Bool, (UInt32,), feature)
-cpu_family() = String(Sys.ARCH)
-
-"""
-    cpu_isa()
-
-Return the [`ISA`](@ref) (instruction set architecture) of the current CPU.
-"""
-function cpu_isa()
-    all_features = last(last(get(ISAs_by_family, cpu_family(), "" => [ISA(Set{UInt32}())]))).features
-    return ISA(Set{UInt32}(feat for feat in all_features if test_cpu_feature(feat)))
-end
-
-end # module CPUID
-
+include("cpuid.jl")
 using .CPUID
 
 # This exists to ease compatibility with old-style Platform objects
@@ -628,28 +564,39 @@ const arch_mapping = Dict(
     "powerpc64le" => "p(ower)?pc64le",
 )
 # Keep this in sync with `CPUID.ISAs_by_family`
+# These are the CPUID side of the microarchitectures targeted by GCC flags in BinaryBuilder.jl
 const arch_march_isa_mapping = let
     function get_set(arch, name)
         all = CPUID.ISAs_by_family[arch]
         return all[findfirst(x -> x.first == name, all)].second
     end
     Dict(
+        "i686" => [
+            "i686" => get_set("i686", "i686"),
+            "prescott" => get_set("i686", "prescott"),
+        ],
         "x86_64" => [
             "x86_64" => get_set("x86_64", "x86_64"),
             "avx" => get_set("x86_64", "sandybridge"),
             "avx2" => get_set("x86_64", "haswell"),
             "avx512" => get_set("x86_64", "skylake_avx512"),
         ],
+        "armv6l" => [
+            "arm1176jzfs" => get_set("armv6l", "arm1176jzfs"),
+        ],
         "armv7l" => [
-            "armv7l" => get_set("arm", "armv7l"),
-            "neon" => get_set("arm", "armv7l_neon"),
-            "vfp4" => get_set("arm", "armv7l_neon_vfp4"),
+            "armv7l" => get_set("armv7l", "armv7l"),
+            "neonvfp4" => get_set("armv7l", "armv7l+neon+vfp4"),
         ],
         "aarch64" => [
-            "armv8" => get_set("aarch64", "armv8.0_a"),
-            "thunderx2" => get_set("aarch64", "armv8.1_a"),
-            "carmel" => get_set("aarch64", "armv8.2_a_crypto"),
+            "armv8_0" => get_set("aarch64", "armv8.0-a"),
+            "armv8_1" => get_set("aarch64", "armv8.1-a"),
+            "armv8_2_crypto" => get_set("aarch64", "armv8.2-a+crypto"),
+            "armv8_4_crypto_sve" => get_set("aarch64", "armv8.4-a+crypto+sve"),
         ],
+        "powerpc64le" => [
+            "power8" => get_set("powerpc64le", "power8"),
+        ]
     )
 end
 const os_mapping = Dict(
